[MCUX-80726] Fix static analysis identified issues in lpi2c driver

+++ drivers/lpi2c/fsl_lpi2c_edma.c
@@ -199,7 +199,7 @@ static uint32_t LPI2C_GenerateCommands(lpi2c_master_edma_handle_t *handle)
             uint32_t subaddressRemaining = xfer->subaddressSize;
             while (0U != subaddressRemaining--)
             {
-                uint8_t subaddressByte = (uint8_t)(xfer->subaddress >> (8U * subaddressRemaining)) & 0xffU;
+                uint8_t subaddressByte = (uint8_t)((xfer->subaddress >> (8U * subaddressRemaining)) & 0xffU);
                 cmd[cmdCount++]        = subaddressByte;
             }
         }
@@ -461,7 +461,7 @@ status_t LPI2C_MasterTransferEDMA(LPI2C_Type *base,
          * $Branch Coverage Justification$
          * $ref fsl_lpi2c_edma_c_ref_3$
          */
-        if(handle->remainingCommand > 0)
+        if(handle->remainingCommand > 0U)
         {
             handle->enableTxReadyFlag = true;
             EDMA_EnableChannelInterrupts(handle->rx->base, handle->rx->channel, (uint32_t)kEDMA_MajorInterruptEnable);
@@ -608,7 +608,7 @@ static void LPI2C_MasterEDMACallback(edma_handle_t *dmaHandle, void *userData, b
 
     if((FSL_FEATURE_LPI2C_HAS_SEPARATE_DMA_RX_TX_REQn(base) == 0) && handle->enableTxReadyFlag == true)
     {
-        LPI2C_MasterEnableInterrupts(handle->base, kLPI2C_MasterTxReadyFlag);
+        LPI2C_MasterEnableInterrupts(handle->base, (uint32_t)kLPI2C_MasterTxReadyFlag);
         handle->enableTxReadyFlag = false;
         return;
     }
@@ -644,9 +644,9 @@ static void LPI2C_MasterTransferEdmaHandleIRQ(LPI2C_Type *base, void *lpi2cMaste
      * $Branch Coverage Justification$
      * $ref fsl_lpi2c_edma_c_ref_3$
      */
-    if((0U != (LPI2C_MasterGetEnabledInterrupts(base) & kLPI2C_MasterTxReadyFlag)) && (0U != (status & (uint32_t)kLPI2C_MasterTxReadyFlag)))
+    if((0U != (LPI2C_MasterGetEnabledInterrupts(base) & (uint32_t)kLPI2C_MasterTxReadyFlag)) && (0U != (status & (uint32_t)kLPI2C_MasterTxReadyFlag)))
     {
-        if(handle->remainingCommand > 0)
+        if(handle->remainingCommand > 0U)
         {
             uint32_t i;
             uint32_t maxTxFifo = (handle->base->PARAM & LPI2C_PARAM_MTXFIFO_MASK) >> LPI2C_PARAM_MTXFIFO_SHIFT;
@@ -658,9 +658,9 @@ static void LPI2C_MasterTransferEdmaHandleIRQ(LPI2C_Type *base, void *lpi2cMaste
             handle->remainingCommand -= i;
             handle->commandIndex += i;
         }
-        if(handle->remainingCommand == 0)
+        if(handle->remainingCommand == 0U)
         {
-            LPI2C_MasterDisableInterrupts(handle->base, kLPI2C_MasterTxReadyFlag);
+            LPI2C_MasterDisableInterrupts(handle->base, (uint32_t)kLPI2C_MasterTxReadyFlag);
         }
     }
     else
